Move Table to common
