+++ cpp/dolfin/common/CMakeLists.txt
@@ -8,6 +8,7 @@ set(HEADERS
   RangedIndexSet.h
   Set.h
   SubSystemsManager.h
+  Table.h
   Timer.h
   TimeLogger.h
   TimeLogManager.h
@@ -24,6 +25,7 @@ set(SOURCES
   init.cpp
   MPI.cpp
   SubSystemsManager.cpp
+  Table.cpp
   Timer.cpp
   TimeLogger.cpp
   TimeLogManager.cpp
