+++ cpp/dolfin/common/MPI.h
@@ -19,7 +19,7 @@
 #define MPICH_IGNORE_CXX_SEEK 1
 #include <mpi.h>
 
-#include <dolfin/log/Table.h>
+#include <dolfin/common/Table.h>
 
 namespace dolfin
 {
