--- cpp/dolfin/log/CMakeLists.txt
@@ -1,8 +1,0 @@
-set(HEADERS
-  dolfin_log.h
-  Table.h
-  PARENT_SCOPE)
-
-set(SOURCES
-  Table.cpp
-  PARENT_SCOPE)
