--- cpp/dolfin/log/dolfin_log.h
@@ -1,17 +1,0 @@
-#pragma once
-
-namespace dolfin
-{
-/*! \namespace dolfin::log
-    \brief Interface to logging tools
-
-    Logging system to log messages and the manage warnings and errors.
-*/
-}
-
-/// DOLFIN log interface
-
-// #include <dolfin/log/LogLevel.h>
-// #include <dolfin/log/LogStream.h>
-#include <dolfin/log/Table.h>
-// #include <dolfin/log/log.h>
